Reorganize files in preparation for moving file  drake/multibody/multibody_tree/math/rotation_matrix.h  to  drake/math/rotation_matrix.h
