+++ math/BUILD.bazel
@@ -55,6 +55,7 @@ drake_cc_library(
         "random_rotation.cc",
         "roll_pitch_yaw.cc",
         "rotation_matrix.cc",
+        "rotation_matrix_utilities.cc",
     ],
     hdrs = [
         "axis_angle.h",
@@ -64,6 +65,7 @@ drake_cc_library(
         "roll_pitch_yaw_not_using_quaternion.h",
         "roll_pitch_yaw_using_quaternion.h",
         "rotation_matrix.h",
+        "rotation_matrix_utilities.h",
     ],
     deps = [
         "//common:essential",
@@ -347,7 +349,7 @@ drake_cc_googletest(
 )
 
 drake_cc_googletest(
-    name = "rotation_matrix_test",
+    name = "rotation_matrix_utilities_test",
     deps = [
         ":geometric_transform",
         "//common/test_utilities:eigen_matrix_compare",
